fix: Better Error Handling of Intrinsics (#1896)
* Better Error Handling of Intrinsics

In HTTP API authorizers, we don't yet handle intrinsics, and an attempt
to use them creates a crash when it tries to hash the input `dict`. This
change starts by wrapping that use case properly so that a helpful error
is raised until intrinsics are supported here.

Note: Does not yet handle `!Sub`, looking at options for that.

* Add "NoValue" Test Case

Should not raise, making that explicit.